+++ src/app/@modal/(.)recipes/[recipeKey]/style.module.scss
@@ -4,4 +4,5 @@
   max-width: $layoutMaxWidth;
   width: 80vw;
   height: 90vh;
+  overflow-y: overlay;
 }
