+++ src/app/components/layout/Nav/style.module.scss
@@ -7,7 +7,7 @@
 .container {
   margin: 0 auto;
   padding: 20px 25px;
-  height: 100px;
+  min-height: 100px;
   max-width: $layoutMaxWidth;
   width: 100%;
   display: grid;
