+++ src/app/recipes/[recipeKey]/Recommend/style.module.scss
@@ -1,0 +1,37 @@
+@use '@/styles/theme.scss' as *;
+
+.recommend {
+  margin-top: 40px;
+}
+
+.title {
+  font-size: 1.2rem;
+  margin-bottom: 20px;
+}
+
+.list {
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  grid-auto-rows: 400px;
+  gap: 24px;
+
+  @media #{$laptop} {
+    grid-template-columns: repeat(3, 1fr);
+    grid-auto-rows: 300px;
+  }
+
+  @media #{$tablet} {
+    grid-template-columns: repeat(2, 1fr);
+    grid-auto-rows: 200px;
+  }
+
+  @media #{$mobile} {
+    grid-template-columns: repeat(1, 1fr);
+  }
+}
+
+.item {
+  width: 100%;
+  height: 100%;
+  display: flex;
+}
