+++ src/app/recipes/[recipeKey]/components/IngredientList/index.tsx
@@ -1,6 +1,6 @@
 import style from './style.module.scss';
 
-import { RecipeDetail } from '@/services/recipes/type';
+import { RecipeDetail } from '@/services/recipe/type';
 
 import { getSearchIngredientLink } from '@/helpers/link';
