+++ src/app/recipes/[recipeKey]/components/Product/index.tsx
@@ -5,7 +5,7 @@ import { joinClassNames } from '@/utils/style';
 import { getSearchProductLink } from '@/helpers/link';
 
 import { Product as ProductType } from '@/services/products/type';
-import { RecipeProduct } from '@/services/recipes/type';
+import { RecipeProduct } from '@/services/recipe/type';
 
 import Chip from '@/components/Chip';
 import Icon, { IconProps } from '@/components/Icon';
