+++ src/app/users/[username]/Cards/Cards.tsx
@@ -2,7 +2,7 @@ import Link from 'next/link';
 
 import style from './Cards.module.scss';
 
-import { RecipeSimple } from '@/services/recipes/type';
+import { RecipeSimple } from '@/services/recipe/type';
 
 import { formatTime } from '@/utils/time';
 import { getLimitedText } from '@/utils/text';
