+++ src/components/Card/style.module.scss
@@ -29,10 +29,7 @@
 
 .img {
   object-fit: cover;
-
-  img {
-    border-radius: 10px;
-  }
+  border-radius: 10px;
 }
 
 .top-overlay {
